[LEMS-2982/answerless-table] add stories
